% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{build_info}
\alias{build_info}
\alias{get_status_travis}
\alias{get_status_appveyor}
\alias{get_code_coverage}
\title{Retrieve build information.}
\usage{
get_status_travis(repo_name, user_name = NULL)

get_status_appveyor(repo_name, user_name = NULL)

get_code_coverage(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns build status from Travis CI and Appveyor, and code coverage from
codecov.
}
\examples{
# get travis ci build status
get_status_travis("dplyr", "tidyverse")

# get appveyor build status
get_status_appveyor("dplyr", "tidyverse")

# get codecov code coverage
get_code_coverage("dplyr", "tidyverse")

}
