% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_appveyor}
\alias{check_appveyor}
\title{Check appveyor build status}
\usage{
check_appveyor(user_name, repo_name)
}
\arguments{
\item{user_name}{Name of the GitHub repository.}

\item{repo_name}{Name of the package.}
}
\description{
Return the latest appveyor build status.
}
\examples{
\dontrun{
check_appveyor("rsquaredacademy", "olsrr")
}

}
