% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_cran_results}
\alias{check_cran_results}
\title{Check CRAN results}
\usage{
check_cran_results(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Return latest CRAN build results.
}
\examples{
\dontrun{
check_cran_results("dplyr")
}

}
