% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\docType{class}
\name{CranPackage}
\alias{CranPackage}
\title{Retrieve package information}
\format{An R6 class.}
\description{
Retrieve package information
}
\section{Usage}{

\preformatted{
myPackage <- CranPackage$new("package_name")
myPackage$get_downloads()
myPackage$get_cran_check_results()
myPackage$get_title()
myPackage$get_description()
myPackage$get_version()
myPackage$get_r_dep()
myPackage$get_imports()
myPackage$get_suggest()
myPackage$get_publish_date()
myPackage$get_license()
myPackage$get_authors()
myPackage$get_maintainer()
myPackage$get_urls()
}
}

\section{Arguments}{

\describe{
  \item{package_name}{Name of the R packge.}
}
}

\section{Details}{


To create \code{CranPackage} objects, you need to use \code{CranPackage$new("package_name")}.

\code{myPackage$get_downloads()} will return the downloads of the package
from the RStudio CRAN mirror for the last day, last week, last month and
total downloads.

\code{myPackage$get_cran_check_results()} will return the CRAN check results of
the package.

\code{myPackage$get_title()} will return the title of the package.

\code{myPackage$get_description()} will return the description of the package.

\code{myPackage$get_version()} will return the version of the package.

\code{myPackage$get_r_dep()} will return the R dependency of the package.

\code{myPackage$get_imports()} will return the R packages imported by the package.

\code{myPackage$get_suggests()} will return the R packages suggested by the package.

\code{myPackage$get_publish_date()} will return the date the package was published on CRAN.

\code{myPackage$get_license()} will return the license under which the package has been released.

\code{myPackage$get_authors()} will return the names of the authors of the package.

\code{myPackage$get_maintainer()} will return the name of the maintainer of the package.

\code{myPackage$get_urls()} will return the URLs associated with the package.
}

\examples{
\dontrun{
myPackage <- CranPackage$new("dplyr")
myPackage$get_title()
myPackage$get_version()
myPackage$get_r_deps()
myPackage$get_imports()
}

}
