% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_code_coverage}
\alias{get_code_coverage}
\title{Code coverage}
\usage{
get_code_coverage(user_name, repo_name)
}
\arguments{
\item{user_name}{Name of the GitHub repository.}

\item{repo_name}{Name of the package.}
}
\description{
Return the code coverage of the package.
}
\examples{
\dontrun{
get_code_coverage("rsquaredacademy", "olsrr")
}

}
