% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_cran_downloads}
\alias{get_cran_downloads}
\title{Downloads}
\usage{
get_cran_downloads(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Package downloads from RStudio CRAN mirror.
}
\examples{
\dontrun{
get_cran_downloads("dplyr")
}

}
