% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_license}
\alias{get_cran_license}
\title{Retrieve license}
\usage{
get_cran_license(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the license type of the package.
}
\examples{
get_cran_license("dplyr")

}
