% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_maintainer}
\alias{get_cran_maintainer}
\title{Retrieve maintainer}
\usage{
get_cran_maintainer(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the details of the maintainer of the package.
}
\examples{
\dontrun{
get_cran_maintainer("dplyr")
}

}
