% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_cran_results}
\alias{get_cran_results}
\title{Check results}
\usage{
get_cran_results(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Return latest CRAN build results.
}
\examples{
\dontrun{
get_cran_results("dplyr")
}

}
