% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_suggests}
\alias{get_cran_suggests}
\title{Retrieve suggests}
\usage{
get_cran_suggests(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the list of packages suggested.
}
\examples{
get_cran_suggests("dplyr")

}
