% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_version}
\alias{get_cran_version}
\title{Retrieve version}
\usage{
get_cran_version(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the latest version of the package from CRAN.
}
\examples{
get_cran_version("dplyr")

}
