% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{get_downloads}
\alias{get_downloads}
\title{Downloads}
\usage{
get_downloads(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Package downloads from cranlogs.
}
\examples{
\dontrun{
get_downloads("dplyr")
}

}
