% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_branches}
\alias{get_gh_branches}
\title{Branches}
\usage{
get_gh_branches(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns names of branches.
}
\examples{
\dontrun{
get_gh_branches("dplyr", "tidyverse")
}

}
