% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_coc}
\alias{get_gh_coc}
\title{Code of conduct}
\usage{
get_gh_coc(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns code of conduct of repository.
}
\examples{
\dontrun{
get_gh_coc("dplyr", "tidyverse")
}

}
