% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_issues}
\alias{get_gh_issues}
\title{Issues}
\usage{
get_gh_issues(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns issues of repository.
}
\examples{
\dontrun{
get_gh_issues("dplyr", "tidyverse")
}

}
