% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_labels}
\alias{get_gh_labels}
\title{Labels}
\usage{
get_gh_labels(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns labels of repository.
}
\examples{
\dontrun{
get_gh_labels("dplyr", "tidyverse")
}


}
