% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_license}
\alias{get_gh_license}
\title{License}
\usage{
get_gh_license(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns license of repository.
}
\examples{
\dontrun{
get_gh_license("dplyr", "tidyverse")
}

}
