% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_milestones}
\alias{get_gh_milestones}
\title{Milestones}
\usage{
get_gh_milestones(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns milestones of repository.
}
\examples{
\dontrun{
get_gh_milestones("dplyr", "tidyverse")
}

}
