% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_pr}
\alias{get_gh_pr}
\title{Pull requests}
\usage{
get_gh_pr(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns pull requests of repository.
}
\examples{
\dontrun{
get_gh_pr("dplyr", "tidyverse")
}

}
