% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_releases}
\alias{get_gh_releases}
\title{Releases}
\usage{
get_gh_releases(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns releases of repository.
}
\examples{
\dontrun{
get_gh_releases("dplyr", "tidyverse")
}

}
