% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_stats}
\alias{get_gh_stats}
\title{Stars, forks and issues}
\usage{
get_gh_stats(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns number of stars, forks and open issues of a GitHub repository.
}
\examples{
\dontrun{
get_gh_stats("dplyr", "tidyverse")
}

}
