% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_username}
\alias{get_gh_username}
\title{GitHub username}
\usage{
get_gh_username(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Returns the GitHub user or organization name.
}
\examples{
\dontrun{
get_gh_username("olsrr")
}

}
