% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_details}
\alias{get_pkg_details}
\title{Package details}
\usage{
get_pkg_details(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Extracts package details from crandb API.
}
\examples{
get_pkg_details("dplyr")

}
