% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_imports}
\alias{get_pkg_imports}
\title{Imports}
\usage{
get_pkg_imports(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the list of packages imported.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_imports()
}

}
