% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_license}
\alias{get_pkg_license}
\title{License}
\usage{
get_pkg_license(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the license type of the package.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_license()
}

}
