% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_maintainer}
\alias{get_pkg_maintainer}
\title{Maintainer}
\usage{
get_pkg_maintainer(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the details of the maintainer of the package.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_maintainer()
}

}
