% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_publish_date}
\alias{get_pkg_publish_date}
\title{Published date}
\usage{
get_pkg_publish_date(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the latest date on which the package was published to CRAN.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_publish_date()
}

}
