% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_suggests}
\alias{get_pkg_suggests}
\title{Suggests}
\usage{
get_pkg_suggests(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the list of packages suggested.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_suggests()
}

}
