% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_urls}
\alias{get_pkg_urls}
\title{URL}
\usage{
get_pkg_urls(pkg_details)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the list of URLs associated with the package.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_urls()
}

}
