% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_version}
\alias{get_pkg_version}
\title{Version}
\usage{
get_pkg_version(pkg_detail)
}
\arguments{
\item{pkg_details}{An object of class \code{pkg_details}.}
}
\description{
Retrieve the latest version of the package from CRAN.
}
\examples{
\dontrun{
"dplyr" \%>\%
  get_pkg_details() \%>\%
  get_pkg_version()
}

}
