% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-stack.R
\name{get_so_questions}
\alias{get_so_questions}
\title{Stack Overflow questions}
\usage{
get_so_questions(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Display package related questions on Stack Overflow.
}
\examples{
\dontrun{
get_so_questions("dplyr")	
}

}
