% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_status_travis}
\alias{get_status_travis}
\title{Travis build status}
\usage{
get_status_travis(user_name, repo_name)
}
\arguments{
\item{user_name}{Name of the GitHub repository.}

\item{repo_name}{Name of the package.}
}
\description{
Return the latest travis build status.
}
\examples{
\dontrun{
get_status_travis("rsquaredacademy", "olsrr")
}

}
