% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{github_info}
\alias{github_info}
\alias{get_gh_username}
\alias{get_gh_stats}
\alias{get_gh_branches}
\alias{get_gh_issues}
\alias{get_gh_labels}
\alias{get_gh_milestones}
\alias{get_gh_coc}
\alias{get_gh_license}
\alias{get_gh_pr}
\alias{get_gh_releases}
\title{Retrieve GitHub information.}
\usage{
get_gh_username(repo_name)

get_gh_stats(repo_name, user_name = NULL)

get_gh_branches(repo_name, user_name = NULL)

get_gh_issues(repo_name, user_name = NULL)

get_gh_labels(repo_name, user_name = NULL)

get_gh_milestones(repo_name, user_name = NULL)

get_gh_coc(repo_name, user_name = NULL)

get_gh_license(repo_name, user_name = NULL)

get_gh_pr(repo_name, user_name = NULL)

get_gh_releases(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Returns the GitHub username, stars, forks, branches, issues, labels, 
milestones, code of conduct, license, pull requests and releases.
}
\examples{
\dontrun{
get_gh_username("olsrr")
get_gh_stats("dplyr", "tidyverse")
get_gh_branches("dplyr", "tidyverse")
get_gh_issues("dplyr", "tidyverse")
get_gh_labels("dplyr", "tidyverse")
get_gh_milestones("dplyr", "tidyverse")
get_gh_coc("dplyr", "tidyverse")
get_gh_license("dplyr", "tidyverse")
get_gh_pr("dplyr", "tidyverse")
get_gh_releases("dplyr", "tidyverse")
}

}
