% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkginfo.R
\docType{package}
\name{pkginfo}
\alias{pkginfo}
\title{\code{pkginfo} package}
\description{
Retrieve pakcage information.
}
\details{
See the README on
\href{https://github.com/rsquaredacademy/pkginfo}{GitHub}
}
