% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloads.R
\name{show_downloads}
\alias{show_downloads}
\title{Downloads}
\usage{
show_downloads(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Package downloads from cranlogs.
}
\examples{
show_downloads("dplyr")

}
