% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{show_github_info}
\alias{show_github_info}
\title{GitHub info}
\usage{
show_github_info(repo_name, package_name)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{package_name}{Name of the package.}
}
\description{
Displays GitHub related information of a package.
}
\examples{
show_github_info("tidyverse", "dplyr")

}
