% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_appveyor}
\alias{check_appveyor}
\title{Check appveyor build status}
\usage{
check_appveyor(repo_name, package_name)
}
\arguments{
\item{repo_name}{Name of the GitHub repository.}

\item{package_name}{Name of the package.}
}
\description{
Return the lates appveyor build status.
}
\examples{
check_appveyor("rsquaredacademy", "olsrr")

}
