% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_coverage}
\alias{check_coverage}
\title{Check code coverage}
\usage{
check_coverage(repo_name, package_name)
}
\arguments{
\item{repo_name}{Name of the GitHub repository.}

\item{package_name}{Name of the package.}
}
\description{
Return the code coverage of the package.
}
\examples{
\dontrun{
check_coverage("rsquaredacademy", "olsrr")
}

}
