% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_travis}
\alias{check_travis}
\title{Check travis build status}
\usage{
check_travis(repo_name, package_name)
}
\arguments{
\item{repo_name}{Name of the GitHub repository.}

\item{package_name}{Name of the package.}
}
\description{
Return the latest travis build status.
}
\examples{
check_travis("rsquaredacademy", "olsrr")

}
