% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_code_coverage}
\alias{get_code_coverage}
\title{Code coverage}
\usage{
get_code_coverage(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Return the code coverage of the package.
}
\examples{
\dontrun{
get_code_coverage("dplyr", "rsquaredacademy")
}

}
