% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_authors}
\alias{get_cran_authors}
\title{Retrieve authors}
\usage{
get_cran_authors(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the list of authors of the package.
}
\examples{
get_cran_authors("dplyr")

}
