% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_desc}
\alias{get_cran_desc}
\title{Retrieve description}
\usage{
get_cran_desc(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the description of the package from CRAN.
}
\examples{
\dontrun{
get_cran_desc("dplyr")
}

}
