% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_imports}
\alias{get_cran_imports}
\title{Retrieve imports}
\usage{
get_cran_imports(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the list of packages imported.
}
\examples{
\dontrun{
get_cran_imports("dplyr")
}

}
