% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_cran_pub_date}
\alias{get_cran_pub_date}
\title{Published date}
\usage{
get_cran_pub_date(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the latest date on which the package was published to CRAN.
}
\examples{
\dontrun{
get_cran_pub_date("dplyr")
}

}
