% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_cran_r_dep}
\alias{get_cran_r_dep}
\title{Dependency}
\usage{
get_cran_r_dep(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the R version on which the package depends.
}
\examples{
\dontrun{
get_cran_r_dep("dplyr")
}

}
