% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_cran_title}
\alias{get_cran_title}
\title{Title}
\usage{
get_cran_title(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the title of the package from CRAN.
}
\examples{
\dontrun{
get_cran_title("dplyr")
}

}
