% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran-info.R
\name{get_cran_urls}
\alias{get_cran_urls}
\title{Retrieve url}
\usage{
get_cran_urls(package_name)
}
\arguments{
\item{package_name}{Name of the R package.}
}
\description{
Retrieve the list of URLs associated with the package.
}
\examples{
\dontrun{
get_cran_urls("dplyr")
}

}
