% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-info.R
\name{get_gh_branches}
\alias{get_gh_branches}
\title{Branches}
\usage{
get_gh_branches(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns names of branches.
}
\examples{
\dontrun{
get_gh_branches("tidyverse", "dplyr")
}

}
