% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-info.R
\name{get_gh_coc}
\alias{get_gh_coc}
\title{Code of conduct}
\usage{
get_gh_coc(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns code of conduct of repository.
}
\examples{
\dontrun{
get_gh_coc("tidyverse", "dplyr")
}

}
