% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_issues}
\alias{get_gh_issues}
\title{Issues}
\usage{
get_gh_issues(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns issues of repository.
}
\examples{
\dontrun{
get_gh_issues("tidyverse", "dplyr")
}

}
