% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-info.R
\name{get_gh_labels}
\alias{get_gh_labels}
\title{Labels}
\usage{
get_gh_labels(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns labels of repository.
}
\examples{
\dontrun{
get_gh_labels("tidyverse", "dplyr")
}


}
