% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_milestones}
\alias{get_gh_milestones}
\title{Milestones}
\usage{
get_gh_milestones(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns milestones of repository.
}
\examples{
\dontrun{
get_gh_milestones("tidyverse", "dplyr")
}

}
