% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-info.R
\name{get_gh_pr}
\alias{get_gh_pr}
\title{Pull requests}
\usage{
get_gh_pr(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns pull requests of repository.
}
\examples{
\dontrun{
get_gh_pr("tidyverse", "dplyr")
}

}
