% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_releases}
\alias{get_gh_releases}
\title{Releases}
\usage{
get_gh_releases(user_name, repo_name)
}
\arguments{
\item{user_name}{User or organization name.}

\item{repo_name}{Name of the repository.}
}
\description{
Returns releases of repository.
}
\examples{
\dontrun{
get_gh_releases("tidyverse", "dplyr")
}

}
