% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_gh_stats}
\alias{get_gh_stats}
\title{Stars, forks and issues}
\usage{
get_gh_stats(user_name, repo_name)
}
\arguments{
\item{user_name}{Name of the repository.}

\item{repo_name}{Name of the package.}
}
\description{
Returns number of stars, forks and open issues of a GitHub repository.
}
\examples{
\dontrun{
get_gh_stats("tidyverse", "dplyr")
}

}
