% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github-info.R
\name{get_github_info}
\alias{get_github_info}
\title{GitHub info}
\usage{
get_github_info(repo_name, package_name)
}
\arguments{
\item{repo_name}{Name of the repository.}

\item{package_name}{Name of the package.}
}
\description{
Displays GitHub related information of a package.
}
\examples{
\dontrun{
get_github_info("tidyverse", "dplyr")
}

}
