% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-cran.R
\name{get_pkg_cran_check_results}
\alias{get_pkg_cran_check_results}
\title{CRAN check results}
\usage{
get_pkg_cran_check_results(package_name)
}
\arguments{
\item{package_name}{Name of the package.}
}
\description{
Return latest CRAN check results.
}
\examples{
# CRAN check results
get_pkg_cran_check_results("dplyr")

}
