% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_status_appveyor}
\alias{get_status_appveyor}
\title{Appveyor build status}
\usage{
get_status_appveyor(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Return the latest appveyor build status.
}
\examples{
\dontrun{
get_status_appveyor("dplyr", "rsquaredacademy")
}

}
