% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-github.R
\name{get_status_travis}
\alias{get_status_travis}
\title{Travis build status}
\usage{
get_status_travis(repo_name, user_name = NULL)
}
\arguments{
\item{repo_name}{Name of the package.}

\item{user_name}{GitHub user or organization name.}
}
\description{
Return the latest travis build status.
}
\examples{
\dontrun{
get_status_travis("dplyr", "rsquaredacademy")
}

}
